requirements.txt
# Automated CE Outcomes Dashboard - V3
# Python 3.10+

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# HTTP Client (for OpenRouter API)
httpx>=0.25.0
aiohttp>=3.9.0

# Database
aiosqlite>=0.19.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0
scikit-learn>=1.3.0  # For deduplication (cosine similarity utilities)

# LLM APIs
openai>=1.0.0  # For embeddings and tagging

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Logging
structlog>=23.2.0

# Progress bars (for deduplication batch processing)
tqdm>=4.66.0

# Document generation (for executive summaries)
python-docx>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.6.0

# Code quality
black>=23.10.0
ruff>=0.1.0
